% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{adla_list}
\alias{adla_list}
\title{List the available accounts.}
\usage{
adla_list(token, subscription_id)
}
\arguments{
\item{token}{token, can be created using \code{\link[AzureOAuth:oauth_token_azure]{AzureOAuth::oauth_token_azure()}} or
\code{\link[AzureOAuth:oauth_service_token_azure]{AzureOAuth::oauth_service_token_azure()}}}

\item{subscription_id}{character, has form \code{"aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee"}}
}
\value{
List
}
\description{
This function does not (yet) support the query parameters.
Perhaps this can be achieved using (\code{...}). Perhaps is can be achieved by
using some sort of OData-query S3 object
}
\examples{
\dontrun{
  library("httr")
  library("AzureOAuth")

  # assumes this is an Azure native app.
  token <- oauth_token_azure(
    tenant_id = "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee",
    application_id = "ffffffff-gggg-hhhh-iiii-jjjjjjjjjjjj",
    name = "foo"
  )

  adla_list(
    token = token,
    subscription_id = "kkkkkkkk-llll-mmmm-nnnn-oooooooooooo",
  )
}
}
\seealso{
\href{https://docs.microsoft.com/en-us/rest/api/datalakestore/account#Account_List}{Azure Datalake Store Account API reference}
}
