% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_local.R
\name{adla_url}
\alias{adla_url}
\title{Base URL}
\usage{
adla_url(subscription_id, resource_group_name = NULL, name = NULL)
}
\arguments{
\item{subscription_id}{character, has form \code{"aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee"}}

\item{resource_group_name}{character}

\item{name}{character}
}
\value{
\code{url} object returned by \code{\link[httr:parse_url]{httr::parse_url()}}
}
\description{
This is a convenience function to return the base URL for the service.
}
\examples{
  adla_url(
    subscription_id = "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee",
    resource_group_name = "foo",
    name = "bar"
  )
}
\keyword{internal}
