% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unpack_response}
\alias{unpack_response}
\title{unpacks response}
\usage{
unpack_response(response, ...)
}
\arguments{
\item{response}{\code{httr} response object}

\item{...}{other arguments passed to \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON}}}
}
\value{
whatever the JSON spits out
}
\description{
unpacks response
}
\keyword{internal}
