% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{url_path}
\alias{url_path}
\alias{url_path_append}
\alias{url_path_prepend}
\title{Path operators for URL.}
\usage{
url_path(url, ...)

url_path_append(url, ...)

url_path_prepend(url, ...)
}
\arguments{
\item{url}{\code{url} object, as returned by \code{\link[httr:parse_url]{httr::parse_url()}}}

\item{...}{arguments to comprise, append, or preped the path}
}
\value{
\code{url} object
}
\description{
Path operators for URL.
}
\keyword{internal}
