% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{url_query}
\alias{url_query}
\alias{url_query_append}
\title{Path operators for URL.}
\usage{
url_query(url, ...)

url_query_append(url, ...)
}
\arguments{
\item{url}{\code{url} object, as returned by \code{\link[httr:parse_url]{httr::parse_url()}}}

\item{...}{arguments to comprise or append the query}
}
\value{
\code{url} object
}
\description{
Path operators for URL.
}
\keyword{internal}
